+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 - Load potentially big text files as `StringLike` to bypass string size limit
+- Add `mvs-stories` app included in the `molstar` NPM package
+  - Use the app in the corresponding example
 
 ## [v4.15.0] - 2025-05-19
 - IHM improvements: