+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
 ## [Unreleased]
+
+## [v4.16.0] - 2025-05-20
 - Load potentially big text files as `StringLike` to bypass string size limit
 - MolViewSpec extension:
   - Load single-state MVS as if it were multi-state with one state